third_party/blink/web_tests/fast/frames/sandboxed-iframe-navigation-top-expected.txt

CONSOLE WARNING: An iframe which has both allow-scripts and allow-same-origin for its sandbox attribute can escape its sandboxing.
PASS
